Renaming the RosalutionToast to ToastDialog to correspond with the other dialog components
